server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["amirgamil <amir.amgad01@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10.1, <4"
langchain = "0.0.173"
openai = "^0.26.5"
nltk = "^3.8.1"
faiss-cpu = "^1.7.3"
bs4 = "^0.0.1"
unstructured = "^0.4.8"
fastapi = "^0.91.0"
uvicorn = "^0.20.0"
python-multipart = "^0.0.5"
poetry-dotenv-plugin = "^0.1.0"
pydantic = "^1.10.4"
tiktoken = "^0.2.0"
mypy = "^1.0.0"
python-dotenv = "^1.0.0"
prisma = "^0.8.2"
asyncio = "^3.4.3"
slack-sdk = "^3.21.3"
matplotlib = "^3.7.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
